Cleaned up code

+++ src/Painter/Renderer.h
@@ -15,7 +15,7 @@ public:
 	~CRenderer();
 	
 	//rgb = 0 - 255
-	void RenderRGBQuad(glm::vec2 _pos, glm::vec2 _size, cvec3 _rgb, int width = 0);
+	void RenderRGBQuad(glm::vec2 _pos, glm::vec2 _size, cvec3 _rgb, float width = 0);
 
 	GLFWwindow* GetWindow() const { return m_pWindow; }
 	glm::ivec2 GetWindowSize() const { return m_sizeWindow; }
